requirements.txt
# ── Backend (Python/FastAPI) ──
fastapi
uvicorn[standard]
python-multipart
pydantic
PyPDF2
python-dotenv
google-generativeai
python-docx

# ── Legacy Streamlit UI (optional) ──
streamlit
